libm: upstream-glibc: Add exp10 support and some glibc dependencies
Summary:
exp10 isn't in the BSDs, so this is an import via upstream glibc.
exp10 is also much faster than pow(10,x). With an additional patch,
our llvm backend will convert pow(10,x) to exp10(x).

Test Plan: this needs MASSIVE cleanup still. Much of this code isn't actually upstream...OR libm so not sure what to do with it yet

Reviewers: randall.rushing, jriordan001, anders3408

Differential Revision: http://review.codefirex.com/D178

Conflicts:
	libm/Android.mk